patch: updated vulnerability design, styling cleanup
