+++ src/components/Tag/Tabs/VulnerabilitiesDetails.jsx
@@ -4,249 +4,69 @@ import React, { useEffect, useMemo, useState, useRef } from 'react';
 import { api, endpoints } from '../../../api';
 
 // components
-import Collapse from '@mui/material/Collapse';
-import { Box, Card, CardContent, Divider, Stack, Typography, InputBase } from '@mui/material';
+import { Stack, Typography, InputBase } from '@mui/material';
 import makeStyles from '@mui/styles/makeStyles';
 import { host } from '../../../host';
 import { debounce, isEmpty } from 'lodash';
-import { Link } from 'react-router-dom';
 import Loading from '../../Shared/Loading';
-import { KeyboardArrowDown, KeyboardArrowRight } from '@mui/icons-material';
-import { VulnerabilityChipCheck } from 'utilities/vulnerabilityAndSignatureCheck';
 import { mapCVEInfo } from 'utilities/objectModels';
-import { CVE_FIXEDIN_PAGE_SIZE, EXPLORE_PAGE_SIZE } from 'utilities/paginationConstants';
+import { EXPLORE_PAGE_SIZE } from 'utilities/paginationConstants';
 import SearchIcon from '@mui/icons-material/Search';
 
-const useStyles = makeStyles(() => ({
-  card: {
-    display: 'flex',
-    flexDirection: 'row',
-    alignItems: 'center',
-    background: '#FFFFFF',
-    boxShadow: '0rem 0.3125rem 0.625rem rgba(131, 131, 131, 0.08)',
-    borderRadius: '1.875rem',
-    flex: 'none',
-    alignSelf: 'stretch',
-    flexGrow: 0,
-    order: 0,
-    width: '100%',
-    marginTop: '2rem',
-    marginBottom: '2rem'
-  },
-  content: {
-    textAlign: 'left',
-    color: '#606060',
-    padding: '2% 3% 2% 3%',
-    width: '100%'
-  },
+import VulnerabilitiyCard from '../../Shared/VulnerabilityCard';
+
+const useStyles = makeStyles((theme) => ({
   title: {
-    color: '#828282',
-    fontSize: '1rem',
-    paddingRight: '0.5rem',
-    paddingBottom: '0.5rem',
-    paddingTop: '0.5rem'
-  },
-  values: {
-    color: '#000000',
-    fontSize: '1rem',
+    color: theme.palette.primary.main,
+    fontSize: '1.5rem',
     fontWeight: '600',
-    paddingBottom: '0.5rem',
-    paddingTop: '0.5rem',
-    textOverflow: 'ellipsis'
+    marginBottom: '0'
   },
-  link: {
-    color: '#52637A',
+  cveId: {
+    color: theme.palette.primary.main,
     fontSize: '1rem',
-    letterSpacing: '0.009375rem',
-    paddingRight: '1rem',
-    textDecorationLine: 'underline'
+    fontWeight: 400,
+    textDecoration: 'underline'
   },
-  monitor: {
-    width: '27.25rem',
-    height: '24.625rem',
-    paddingTop: '2rem'
+  cveSummary: {
+    color: theme.palette.secondary.dark,
+    fontSize: '0.75rem',
+    fontWeight: '600',
+    textOverflow: 'ellipsis',
+    marginTop: '0.5rem'
   },
   none: {
     color: '#52637A',
     fontSize: '1.4rem',
     fontWeight: '600'
   },
-  dropdown: {
-    flexDirection: 'row',
-    alignItems: 'center'
-  },
-  dropdownText: {
-    color: '#1479FF',
-    paddingTop: '1rem',
-    fontSize: '1rem',
-    fontWeight: '600',
-    cursor: 'pointer',
-    textAlign: 'center'
-  },
   search: {
     position: 'relative',
-    minWidth: '100%',
+    maxWidth: '100%',
     flexDirection: 'row',
-    marginBottom: '1.7rem',
-    boxShadow: '0rem 0.3125rem 0.625rem rgba(131, 131, 131, 0.08)',
-    border: '0.125rem solid #E7E7E7',
-    borderRadius: '1rem',
-    zIndex: 1155
+    alignItems: 'center',
+    justifyContent: 'space-between',
+    boxShadow: 'none',
+    border: '0.063rem solid #E7E7E7',
+    borderRadius: '0.625rem'
   },
   searchIcon: {
     color: '#52637A',
     paddingRight: '3%'
   },
+  searchInputBase: {
+    width: '90%',
+    paddingLeft: '1.5rem',
+    height: 40,
+    color: 'rgba(0, 0, 0, 0.6)'
+  },
   input: {
     color: '#464141',
-    marginLeft: 1,
-    width: '90%'
+    '&::placeholder': {
+      opacity: '1'
+    }
   }
 }));
-
-function VulnerabilitiyCard(props) {
-  const classes = useStyles();
-  const { cve, name } = props;
-  const [openDesc, setOpenDesc] = useState(false);
-  const [openFixed, setOpenFixed] = useState(false);
-  const [loadingFixed, setLoadingFixed] = useState(true);
-  const [fixedInfo, setFixedInfo] = useState([]);
-  const abortController = useMemo(() => new AbortController(), []);
-
-  // pagination props
-  const [pageNumber, setPageNumber] = useState(1);
-  const [isEndOfList, setIsEndOfList] = useState(false);
-
-  const getPaginatedResults = () => {
-    if (!openFixed || isEndOfList) {
-      return;
-    }
-    setLoadingFixed(true);
-    api
-      .get(
-        `${host()}${endpoints.imageListWithCVEFixed(cve.id, name, { pageNumber, pageSize: CVE_FIXEDIN_PAGE_SIZE })}`,
-        abortController.signal
-      )
-      .then((response) => {
-        if (response.data && response.data.data) {
-          const fixedTagsList = response.data.data.ImageListWithCVEFixed?.Results?.map((e) => e.Tag);
-          setFixedInfo((previousState) => [...previousState, ...fixedTagsList]);
-          setIsEndOfList(
-            [...fixedInfo, ...fixedTagsList].length >= response.data.data.ImageListWithCVEFixed?.Page?.TotalCount
-          );
-        }
-        setLoadingFixed(false);
-      })
-      .catch((e) => {
-        console.error(e);
-        setIsEndOfList(true);
-        setLoadingFixed(false);
-      });
-  };
-
-  useEffect(() => {
-    getPaginatedResults();
-    return () => {
-      abortController.abort();
-    };
-  }, [openFixed, pageNumber]);
-
-  const loadMore = () => {
-    if (loadingFixed || isEndOfList) return;
-    setPageNumber((pageNumber) => pageNumber + 1);
-  };
-
-  const renderFixedVer = () => {
-    if (!isEmpty(fixedInfo)) {
-      return fixedInfo.map((tag, index) => {
-        return (
-          <Link key={index} to={`/image/${encodeURIComponent(name)}/tag/${tag}`} className={classes.link}>
-            {tag}
-          </Link>
-        );
-      });
-    } else {
-      return 'Not fixed';
-    }
-  };
-
-  const renderLoadMore = () => {
-    return (
-      !isEndOfList && (
-        <Typography
-          sx={{
-            color: '#3366CC',
-            cursor: 'pointer',
-            fontSize: '1rem',
-            letterSpacing: '0.009375rem',
-            paddingRight: '1rem',
-            textDecorationLine: 'underline'
-          }}
-          onClick={loadMore}
-          component="div"
-        >
-          Load more
-        </Typography>
-      )
-    );
-  };
-
-  return (
-    <Card className={classes.card} raised>
-      <CardContent className={classes.content}>
-        <Stack sx={{ flexDirection: 'row' }}>
-          <Typography variant="body1" align="left" className={classes.values}>
-            {' '}
-            {cve.id}
-          </Typography>
-        </Stack>
-        <VulnerabilityChipCheck vulnerabilitySeverity={cve.severity} />
-        <Stack sx={{ flexDirection: 'row' }}>
-          <Typography variant="body1" align="left" className={classes.values}>
-            {' '}
-            {cve.title}
-          </Typography>
-        </Stack>
-        <Stack className={classes.dropdown} onClick={() => setOpenFixed(!openFixed)}>
-          {!openFixed ? (
-            <KeyboardArrowRight className={classes.dropdownText} />
-          ) : (
-            <KeyboardArrowDown className={classes.dropdownText} />
-          )}
-          <Typography className={classes.dropdownText}>Fixed in</Typography>
-        </Stack>
-        <Collapse in={openFixed} timeout="auto" unmountOnExit>
-          <Box sx={{ width: '100%' }}>
-            {loadingFixed ? (
-              'Loading...'
-            ) : (
-              <Stack direction="row" sx={{ flexWrap: 'wrap' }}>
-                {renderFixedVer()}
-                {renderLoadMore()}
-              </Stack>
-            )}
-          </Box>
-        </Collapse>
-        <Stack className={classes.dropdown} onClick={() => setOpenDesc(!openDesc)}>
-          {!openDesc ? (
-            <KeyboardArrowRight className={classes.dropdownText} />
-          ) : (
-            <KeyboardArrowDown className={classes.dropdownText} />
-          )}
-          <Typography className={classes.dropdownText}>Description</Typography>
-        </Stack>
-        <Collapse in={openDesc} timeout="auto" unmountOnExit>
-          <Box>
-            <Typography variant="body2" align="left" sx={{ color: '#0F2139', fontSize: '1rem' }}>
-              {' '}
-              {cve.description}{' '}
-            </Typography>
-          </Box>
-        </Collapse>
-      </CardContent>
-    </Card>
-  );
-}
 
 function VulnerabilitiesDetails(props) {
   const classes = useStyles();
@@ -369,48 +189,23 @@ function VulnerabilitiesDetails(props) {
   };
 
   return (
-    <div data-testid="vulnerability-container">
-      <Typography
-        variant="h4"
-        gutterBottom
-        component="div"
-        align="left"
-        style={{
-          color: 'rgba(0, 0, 0, 0.87)',
-          fontSize: '1.5rem',
-          fontWeight: '600',
-          paddingTop: '0.5rem'
-        }}
-      >
+    <Stack direction="column" spacing="1rem" data-testid="vulnerability-container">
+      <Typography variant="h4" gutterBottom component="div" align="left" className={classes.title}>
         Vulnerabilities
       </Typography>
-      <Divider
-        variant="fullWidth"
-        sx={{
-          margin: '5% 0% 5% 0%',
-          background: 'rgba(0, 0, 0, 0.38)',
-          height: '0.00625rem',
-          width: '100%'
-        }}
-      />
-      <Stack className={classes.search} direction="row" alignItems="center" justifyContent="space-between" spacing={2}>
+      <Stack className={classes.search}>
         <InputBase
-          style={{ paddingLeft: 10, height: 46, color: 'rgba(0, 0, 0, 0.6)' }}
-          placeholder={'Search for Vulnerabilities...'}
-          className={classes.input}
+          placeholder={'Search'}
+          classes={{ root: classes.searchInputBase, input: classes.input }}
           onChange={debouncedChangeHandler}
         />
         <div className={classes.searchIcon}>
           <SearchIcon />
         </div>
       </Stack>
-      <Stack direction="column" spacing={2}>
-        <Stack direction="column" spacing={2}>
-          {renderCVEs()}
-          {renderListBottom()}
-        </Stack>
-      </Stack>
-    </div>
+      {renderCVEs()}
+      {renderListBottom()}
+    </Stack>
   );
 }
